[210706] ADD ApplicationTest UPDATE MyGalleryProject

+++ MyGalleryProject/src/main/java/com/koreait/mygallery/dto/Board.java
@@ -14,7 +14,7 @@ public class Board {
 	private String postdate;
 	private String ip;
 	private int state;
-	private int groupno;
+	private long groupno;
 	private int groupord;
 	private int depth;
 }
